fix: 🐛 updates as PR comments
